requirements.txt
Flask==3.0.0
boto3==1.34.0
cryptography==41.0.5
python-dotenv==1.0.0
werkzeug==3.0.1
